+++ src/PortalBase/Builder/Component/Receiver.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Portal\Base\Builder\Component;
 
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString;
 use Heptacom\HeptaConnect\Dataset\Base\TypedDatasetEntityCollection;
 use Heptacom\HeptaConnect\Portal\Base\Builder\BindThisTrait;
 use Heptacom\HeptaConnect\Portal\Base\Builder\ResolveArgumentsTrait;
@@ -19,10 +20,7 @@ final class Receiver extends ReceiverContract
     use BindThisTrait;
     use ResolveArgumentsTrait;
 
-    /**
-     * @var class-string<DatasetEntityContract>
-     */
-    private string $type;
+    private EntityTypeClassString $entityType;
 
     private ?SerializableClosure $batchMethod;
 
@@ -33,7 +31,7 @@ final class Receiver extends ReceiverContract
         $batch = $token->getBatch();
         $run = $token->getRun();
 
-        $this->type = $token->getType();
+        $this->entityType = $token->getEntityType();
         $this->batchMethod = $batch instanceof \Closure ? new SerializableClosure($batch) : null;
         $this->runMethod = $run instanceof \Closure ? new SerializableClosure($run) : null;
     }
@@ -50,7 +48,7 @@ final class Receiver extends ReceiverContract
 
     public function supports(): string
     {
-        return $this->type;
+        return $this->entityType->getClassString();
     }
 
     protected function batch(
@@ -92,7 +90,7 @@ final class Receiver extends ReceiverContract
                 ?string $propertyType,
                 ContainerInterface $container
             ) use ($entity) {
-                if (\is_string($propertyType) && \is_a($propertyType, $this->supports(), true)) {
+                if (\is_string($propertyType) && $this->getSupportedEntityType()->matchClassStringIsOfType($propertyType)) {
                     return $entity;
                 }