+++ src/PortalBase/Emission/Contract/EmitterContract.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Portal\Base\Emission\Contract;
 
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\Exception\InvalidClassNameException;
+use Heptacom\HeptaConnect\Dataset\Base\Exception\InvalidSubtypeClassNameException;
+use Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Exception\UnsupportedDatasetEntityException;
 use Psr\Log\LoggerInterface;
 
@@ -18,6 +21,8 @@ abstract class EmitterContract
     private ?string $runDeclaringClass = null;
 
     private ?string $batchDeclaringClass = null;
+
+    private ?EntityTypeClassString $supportedEntityType = null;
 
     /**
      * First entrypoint to handle an emission in this flow component.
@@ -41,11 +46,22 @@ abstract class EmitterContract
     }
 
     /**
+     * Returns the supported entity type.
+     *
+     * @throws InvalidClassNameException
+     * @throws InvalidSubtypeClassNameException
+     */
+    final public function getSupportedEntityType(): EntityTypeClassString
+    {
+        return $this->supportedEntityType ??= new EntityTypeClassString($this->supports());
+    }
+
+    /**
      * Must return the supported entity type.
      *
      * @return class-string<DatasetEntityContract>
      */
-    abstract public function supports(): string;
+    abstract protected function supports(): string;
 
     /**
      * The entrypoint for handling an emission with the least need of additional programming.
@@ -67,11 +83,7 @@ abstract class EmitterContract
             return false;
         }
 
-        if (!\is_a($entity, $this->supports(), false)) {
-            return false;
-        }
-
-        return true;
+        return $this->getSupportedEntityType()->matchObjectIsOfType($entity);
     }
 
     /**
@@ -113,8 +125,8 @@ abstract class EmitterContract
                         \sprintf(
                             'Emitter "%s" returned object of unsupported type. Expected "%s" but got "%s".',
                             static::class,
-                            $this->supports(),
-                            \get_class($entity)
+                            $this->getSupportedEntityType(),
+                            $entity::class()
                         )
                     ));
 
@@ -174,8 +186,8 @@ abstract class EmitterContract
                     $logger->error(\sprintf(
                         'Emitter "%s" returned object of unsupported type. Expected "%s" but got "%s".',
                         static::class,
-                        $this->supports(),
-                        $entity === null ? 'null' : \get_class($entity)
+                        $this->getSupportedEntityType(),
+                        $entity === null ? 'null' : $entity::class()
                     ));
                 }
             } elseif ($entity instanceof DatasetEntityContract) {