+++ src/PortalBase/Exploration/Contract/ExplorerContract.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Portal\Base\Exploration\Contract;
 
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\Exception\InvalidClassNameException;
+use Heptacom\HeptaConnect\Dataset\Base\Exception\InvalidSubtypeClassNameException;
+use Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Exception\UnsupportedDatasetEntityException;
 use Psr\Log\LoggerInterface;
 
@@ -15,6 +18,8 @@ use Psr\Log\LoggerInterface;
  */
 abstract class ExplorerContract
 {
+    private ?EntityTypeClassString $supportedEntityType = null;
+
     /**
      * First entrypoint to handle an exploration in this flow component.
      * It allows direct stack handling manipulation. @see ExplorerStackInterface
@@ -29,11 +34,22 @@ abstract class ExplorerContract
     }
 
     /**
+     * Returns the supported entity type.
+     *
+     * @throws InvalidClassNameException
+     * @throws InvalidSubtypeClassNameException
+     */
+    final public function getSupportedEntityType(): EntityTypeClassString
+    {
+        return $this->supportedEntityType ??= new EntityTypeClassString($this->supports());
+    }
+
+    /**
      * Must return the supported entity type.
      *
      * @return class-string<DatasetEntityContract>
      */
-    abstract public function supports(): string;
+    abstract protected function supports(): string;
 
     /**
      * The entrypoint for handling an exploration with the least need of additional programming.
@@ -59,7 +75,7 @@ abstract class ExplorerContract
             return false;
         }
 
-        return \is_string($entity) || \is_int($entity) || \is_a($entity, $this->supports(), false);
+        return \is_string($entity) || \is_int($entity) || $this->getSupportedEntityType()->matchObjectIsOfType($entity);
     }
 
     /**