+++ src/PortalBase/Exploration/ExplorerCollection.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Portal\Base\Exploration;
 
-use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
 use Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString;
 use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExplorerContract;
 
 /**
@@ -14,13 +14,13 @@ use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExplorerContract;
 class ExplorerCollection extends AbstractObjectCollection
 {
     /**
-     * @param class-string<DatasetEntityContract> $entityType
-     *
      * @return iterable<int, ExplorerContract>
      */
-    public function bySupport(string $entityType): iterable
+    public function bySupport(EntityTypeClassString $entityType): iterable
     {
-        return $this->filter(fn (ExplorerContract $explorer) => $entityType === $explorer->supports());
+        return $this->filter(
+            static fn (ExplorerContract $explorer): bool => $entityType->same($explorer->getSupportedEntityType())
+        );
     }
 
     /**