+++ src/PortalBase/Mapping/MappingComponentStruct.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Portal\Base\Mapping;
 
-use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString;
 use Heptacom\HeptaConnect\Portal\Base\Mapping\Contract\MappingComponentStructContract;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 
@@ -12,18 +12,15 @@ final class MappingComponentStruct extends MappingComponentStructContract
 {
     protected PortalNodeKeyInterface $portalNodeKey;
 
-    /**
-     * @psalm-var class-string<DatasetEntityContract>
-     */
-    protected string $entityType;
+    protected EntityTypeClassString $entityType;
 
     protected string $externalId;
 
-    /**
-     * @psalm-param class-string<DatasetEntityContract> $entityType
-     */
-    public function __construct(PortalNodeKeyInterface $portalNodeKey, string $entityType, string $externalId)
-    {
+    public function __construct(
+        PortalNodeKeyInterface $portalNodeKey,
+        EntityTypeClassString $entityType,
+        string $externalId
+    ) {
         $this->portalNodeKey = $portalNodeKey;
         $this->entityType = $entityType;
         $this->externalId = $externalId;
@@ -34,10 +31,7 @@ final class MappingComponentStruct extends MappingComponentStructContract
         return $this->portalNodeKey;
     }
 
-    /**
-     * @psalm-return class-string<DatasetEntityContract>
-     */
-    public function getEntityType(): string
+    public function getEntityType(): EntityTypeClassString
     {
         return $this->entityType;
     }