+++ src/PortalBase/Reception/Contract/ReceiverContract.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Portal\Base\Reception\Contract;
 
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\Exception\InvalidClassNameException;
+use Heptacom\HeptaConnect\Dataset\Base\Exception\InvalidSubtypeClassNameException;
+use Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString;
 use Heptacom\HeptaConnect\Dataset\Base\TypedDatasetEntityCollection;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Exception\UnsupportedDatasetEntityException;
 
@@ -13,6 +16,8 @@ use Heptacom\HeptaConnect\Portal\Base\Portal\Exception\UnsupportedDatasetEntityE
  */
 abstract class ReceiverContract
 {
+    private ?EntityTypeClassString $supportedEntityType = null;
+
     /**
      * First entrypoint to handle a reception in this flow component.
      * It allows direct stack handling manipulation. @see ReceiverStackInterface
@@ -30,11 +35,22 @@ abstract class ReceiverContract
     }
 
     /**
+     * Returns the supported entity type.
+     *
+     * @throws InvalidClassNameException
+     * @throws InvalidSubtypeClassNameException
+     */
+    final public function getSupportedEntityType(): EntityTypeClassString
+    {
+        return $this->supportedEntityType ??= new EntityTypeClassString($this->supports());
+    }
+
+    /**
      * Must return the supported entity type.
      *
      * @return class-string<DatasetEntityContract>
      */
-    abstract public function supports(): string;
+    abstract protected function supports(): string;
 
     /**
      * The entrypoint for handling a reception with the least need of additional programming.
@@ -69,7 +85,7 @@ abstract class ReceiverContract
      */
     final protected function isSupported(DatasetEntityContract $entity): bool
     {
-        return \is_a($entity, $this->supports(), false);
+        return $this->getSupportedEntityType()->matchObjectIsOfType($entity);
     }
 
     /**
@@ -96,7 +112,7 @@ abstract class ReceiverContract
         TypedDatasetEntityCollection $entities,
         ReceiveContextInterface $context
     ): iterable {
-        if (!\is_a($entities->getType(), $this->supports(), true)) {
+        if (!$entities->getEntityType()->same($this->getSupportedEntityType())) {
             foreach ($entities as $entity) {
                 $context->markAsFailed($entity, new UnsupportedDatasetEntityException());
             }
@@ -120,7 +136,7 @@ abstract class ReceiverContract
         TypedDatasetEntityCollection $entities,
         ReceiveContextInterface $context
     ): iterable {
-        if (!\is_a($entities->getType(), $this->supports(), true)) {
+        if (!$entities->getEntityType()->same($this->getSupportedEntityType())) {
             foreach ($entities as $entity) {
                 $context->markAsFailed($entity, new UnsupportedDatasetEntityException());
             }