+++ src/PortalBase/Reception/ReceiverCollection.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Portal\Base\Reception;
 
-use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
 use Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString;
 use Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiverContract;
 
 /**
@@ -14,13 +14,13 @@ use Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiverContract;
 class ReceiverCollection extends AbstractObjectCollection
 {
     /**
-     * @param class-string<DatasetEntityContract> $entityType
-     *
      * @return iterable<int, ReceiverContract>
      */
-    public function bySupport(string $entityType): iterable
+    public function bySupport(EntityTypeClassString $entityType): iterable
     {
-        return $this->filter(fn (ReceiverContract $emitter) => $entityType === $emitter->supports());
+        return $this->filter(
+            static fn (ReceiverContract $emitter): bool => $entityType->same($emitter->getSupportedEntityType())
+        );
     }
 
     /**