+++ test/PortalBase/Emission/EmitterCollectionTest.php
@@ -11,9 +11,15 @@ use Heptacom\HeptaConnect\Portal\Base\Test\Fixture\SecondEntity;
 use PHPUnit\Framework\TestCase;
 
 /**
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringReferenceContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\SubtypeClassStringContract
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString
  * @covers \Heptacom\HeptaConnect\Portal\Base\Emission\EmitterCollection
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Emission\Contract\EmitterContract
  */
 final class EmitterCollectionTest extends TestCase
 {
@@ -28,10 +34,10 @@ final class EmitterCollectionTest extends TestCase
             $this->getEmitter(SecondEntity::class),
             $this->getEmitter(FirstEntity::class),
         ]);
-        static::assertNotEmpty($collection->bySupport(FirstEntity::class));
-        static::assertNotEmpty($collection->bySupport(SecondEntity::class));
-        static::assertCount(3, $collection->bySupport(FirstEntity::class));
-        static::assertCount(2, $collection->bySupport(SecondEntity::class));
+        static::assertNotEmpty($collection->bySupport(FirstEntity::class()));
+        static::assertNotEmpty($collection->bySupport(SecondEntity::class()));
+        static::assertCount(3, $collection->bySupport(FirstEntity::class()));
+        static::assertCount(2, $collection->bySupport(SecondEntity::class()));
     }
 
     private function getEmitter(string $support): EmitterContract