+++ test/PortalBase/Reception/ReceiverStackTest.php
@@ -4,20 +4,25 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Portal\Base\Test\Reception;
 
-use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Core\Test\Fixture\FooBarEntity;
 use Heptacom\HeptaConnect\Dataset\Base\TypedDatasetEntityCollection;
 use Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiveContextInterface;
 use Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiverContract;
 use Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiverStackInterface;
 use Heptacom\HeptaConnect\Portal\Base\Reception\ReceiverStack;
+use Heptacom\HeptaConnect\Portal\Base\Test\Fixture\FirstEntity;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @covers \Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringReferenceContract
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\SubtypeClassStringContract
  * @covers \Heptacom\HeptaConnect\Dataset\Base\DatasetEntityCollection
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\EntityTypeClassString
  * @covers \Heptacom\HeptaConnect\Dataset\Base\TypedDatasetEntityCollection
  * @covers \Heptacom\HeptaConnect\Portal\Base\Reception\ReceiverStack
  */
@@ -27,16 +32,16 @@ final class ReceiverStackTest extends TestCase
     {
         $stack = new ReceiverStack([]);
         static::assertCount(0, $stack->next(
-            new TypedDatasetEntityCollection(DatasetEntityContract::class),
+            new TypedDatasetEntityCollection(FirstEntity::class()),
             $this->createMock(ReceiveContextInterface::class)
         ));
     }
 
     public function testStackCallsEveryone(): void
     {
-        $result1 = $this->createMock(DatasetEntityContract::class);
-        $result2 = $this->createMock(DatasetEntityContract::class);
-        $result3 = $this->createMock(DatasetEntityContract::class);
+        $result1 = new FooBarEntity();
+        $result2 = new FooBarEntity();
+        $result3 = new FooBarEntity();
 
         $receiver1 = $this->createMock(ReceiverContract::class);
         $receiver1->expects(static::once())
@@ -64,6 +69,6 @@ final class ReceiverStackTest extends TestCase
             });
 
         $stack = new ReceiverStack([$receiver1, $receiver2, $receiver3]);
-        static::assertCount(3, $stack->next(new TypedDatasetEntityCollection(DatasetEntityContract::class), $this->createMock(ReceiveContextInterface::class)));
+        static::assertCount(3, $stack->next(new TypedDatasetEntityCollection(FooBarEntity::class()), $this->createMock(ReceiveContextInterface::class)));
     }
 }